% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\title{Mean of a sparse_numeric vector}
\usage{
\S4method{mean}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} vector.}

\item{...}{Additional arguments (ignored).}
}
\value{
A numeric scalar giving mean.
}
\description{
Computes arithmetic mean of \code{sparse_numeric} vector,
including implicit zeros.
}
