% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric,ANY-method}
\title{Euclidean norm of a sparse_numeric vector}
\usage{
norm(x, type, ...)

\S4method{norm}{sparse_numeric,ANY}(x, type, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} vector.}

\item{type}{Ignored; included for compatibility with \code{\link[base]{norm}}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Numeric scalar giving Euclidean norm.
}
\description{
Computes Euclidean (L2) norm \eqn{sqrt(sum(x^2))} of a
\code{sparse_numeric} vector.
}
