% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\title{Sparse crossproduct (dot product)}
\usage{
sparse_crossprod(x, y, ...)
}
\arguments{
\item{x, y}{Objects of class \code{sparse_numeric} with the same length.}

\item{...}{Additional arguments (ignored).}
}
\value{
Numeric scalar giving the dot product.
}
\description{
Sparse crossproduct (dot product)
}
