% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_numeric-arith-methods}
\alias{sparse_numeric-arith-methods}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Methods for sparse_numeric arithmetic generics}
\usage{
\S4method{sparse_add}{sparse_numeric,sparse_numeric}(x, y, ...)

\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y, ...)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y, ...)

\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} vector.}

\item{y}{A \code{sparse_numeric} vector.}

\item{...}{Additional arguments (ignored).}
}
\value{
A \code{sparse_numeric} vector, except for \code{sparse_crossprod},
which returns a numeric scalar.
}
\description{
These methods implement the actual arithmetic operations for
\code{sparse_numeric} vectors, corresponding to the exported generics
\code{sparse_add}, \code{sparse_sub}, \code{sparse_mult}, and
\code{sparse_crossprod}.
}
