% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_numeric-arith}
\alias{sparse_numeric-arith}
\alias{+,sparse_numeric,sparse_numeric-method}
\alias{-,sparse_numeric,sparse_numeric-method}
\alias{*,sparse_numeric,sparse_numeric-method}
\title{Arithmetic operations for sparse_numeric vectors}
\usage{
\S4method{+}{sparse_numeric,sparse_numeric}(e1, e2)

\S4method{-}{sparse_numeric,sparse_numeric}(e1, e2)

\S4method{*}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{e1}{A \code{sparse_numeric} vector.}

\item{e2}{A \code{sparse_numeric} vector.}
}
\value{
A \code{sparse_numeric} vector.
}
\description{
Implements elementwise arithmetic for two \code{sparse_numeric}
vectors of the same length.
}
