% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\title{Sparse numeric vector}
\description{
S4 class representing a sparse numeric vector.
}
\section{Slots}{

\describe{
\item{\code{value}}{Numeric vector of non-zero entries.}

\item{\code{pos}}{Integer vector of 1-based positions of non-zero entries.}

\item{\code{length}}{Integer scalar giving the full length of the dense vector.}
}}

