% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_numeric-length}
\alias{sparse_numeric-length}
\alias{length,sparse_numeric-method}
\title{Length of a sparse_numeric vector}
\usage{
\S4method{length}{sparse_numeric}(x)
}
\arguments{
\item{x}{A \code{sparse_numeric} vector.}
}
\value{
An integer scalar giving the length.
}
\description{
Returns the full logical length of the underlying dense vector
represented by a \code{sparse_numeric} object.
}
