% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_sub}
\alias{sparse_sub}
\title{Subtract two sparse_numeric vectors}
\usage{
sparse_sub(x, y, ...)
}
\arguments{
\item{x, y}{Objects of class \code{sparse_numeric} with the same length.}

\item{...}{Additional arguments (ignored).}
}
\value{
A \code{sparse_numeric} vector.
}
\description{
Subtract two sparse_numeric vectors
}
